Trabajos de enfoque/Big Data Aplicado/requirements.txt
# ──────────────────────────────────────────────────
#  SmartManuTech — Dependencias Python
# ──────────────────────────────────────────────────

# Mensajería Kafka
kafka-python==2.0.2

# Base de datos Cassandra
cassandra-driver==3.28.0

# Base de datos InfluxDB (series temporales)
influxdb-client==1.36.1

# API REST
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Machine Learning
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.3
joblib==1.3.2
